feat(CouponService): CouponService 생성 및 검증 메서드 이동
