+++ src/main/java/coupon/coupon/domain/CouponRepository.java
@@ -1,0 +1,7 @@
+package coupon.coupon.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
+
+}
