+++ src/test/java/coupon/coupon/domain/CouponTest.java
@@ -10,6 +10,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
+@DisplayName("쿠폰 테스트")
 class CouponTest {
 
     @DisplayName("적절한 이름으로 쿠폰을 생성한다.")
@@ -53,93 +54,6 @@ class CouponTest {
                         1L, tooLongName,
                         1000, 1000, Category.ELECTRONICS,
                         LocalDateTime.now(), LocalDateTime.now().plusDays(30)
-                )
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("적절한 할인금액으로 쿠폰을 생성한다.")
-    @ValueSource(ints = {1000, 1500, 10000})
-    @ParameterizedTest
-    void couponWithValidDiscountAmount(int discountAmount) {
-        assertThatCode(() -> new Coupon(
-                1L, "name",
-                discountAmount, 5000, Category.ELECTRONICS,
-                LocalDateTime.now(), LocalDateTime.now().plusDays(30)
-        )).doesNotThrowAnyException();
-    }
-
-    @DisplayName("할인금액이 최소단위가 아닐 시 예외가 발생한다.")
-    @ValueSource(ints = {501, 1001, 1001})
-    @ParameterizedTest
-    void exception_When_DiscountAmountIsNotMinUnit(int discountAmount) {
-        // when & then
-        assertThatThrownBy(
-                () -> new Coupon(
-                        1L, "test",
-                        discountAmount, 1000, Category.ELECTRONICS,
-                        LocalDateTime.now(), LocalDateTime.now().plusDays(30)
-                )
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("할인 금액이 최소 할인금액 미만일 경우 예외가 발생한다.")
-    @Test
-    void exception_When_DiscountAmountIsLessThanMin() {
-        // given
-        int discount = 999;
-
-        // when & then
-        assertThatThrownBy(
-                () -> new Coupon(
-                        1L, "test",
-                        discount, 1000, Category.ELECTRONICS,
-                        LocalDateTime.now(), LocalDateTime.now().plusDays(30)
-                )
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("할인 금액이 최대 할인금액 초과일 경우 예외가 발생한다.")
-    @Test
-    void exception_When_DiscountAmountIsBiggerThanMax() {
-        // given
-        int discount = 10001;
-
-        // when & then
-        assertThatThrownBy(
-                () -> new Coupon(
-                        1L, "test",
-                        discount, 1000, Category.ELECTRONICS,
-                        LocalDateTime.now(), LocalDateTime.now().plusDays(30)
-                )
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
-    @DisplayName("최소 주문 금액이 최소치를 넘지 못하면 예외가 발생한다.")
-    @ValueSource(ints = {500, 4500, 3000})
-    @ParameterizedTest
-    void exception_When_MinOrderAmountIsLessThanMin(int minOrderAmount) {
-        // when & then
-        assertThatThrownBy(
-                () -> new Coupon(
-                        1L, "test",
-                        1000, minOrderAmount, Category.ELECTRONICS,
-                        LocalDateTime.now(), LocalDateTime.now().plusDays(30)
-                ));
-    }
-
-    @DisplayName("시작일이 종료일보다 이후일 경우 예외가 발생한다.")
-    @Test
-    void exception_When_StartDateIsAfterEndDate() {
-        // given
-        LocalDateTime endDate = LocalDateTime.now();
-        LocalDateTime startDate = endDate.plusDays(1);
-
-        // when & then
-        assertThatThrownBy(
-                () -> new Coupon(
-                        1L, "test",
-                        1000, 5000, Category.ELECTRONICS,
-                        startDate, endDate
                 )
         ).isInstanceOf(IllegalArgumentException.class);
     }
